// types.d.ts
export interface Post {
  id: string;
  title: string;
  body: string;
  date: Date;
  excerpt: string;
  tags: string
}

export interface Project {
  id: string;
  title: string;
  image: string;
  body: string;
  date: Date;
  excerpt: string;
}

export interface P {
  id: string;
  title: string;
  body: string;
  excerpt: string;
}